Bump Swift version to 4.1
