+++ .travis.yml
@@ -1,13 +1,8 @@
-osx_image: xcode8.3
+osx_image: xcode9.3
 language: generic
 matrix:
   include:
     - os: osx
-      env:
-        - PLATFORM=macos
-        - XCODE_ACTION="build-for-testing test-without-building"
-    - os: osx
-      osx_image: xcode9.3
       env:
         - PLATFORM=macos
         - XCODE_ACTION="build-for-testing test-without-building"
@@ -23,15 +18,8 @@ matrix:
       env:
         - PLATFORM=swiftpm
     - os: osx
-      osx_image: xcode9.3
-      env:
-        - PLATFORM=swiftpm
-    - os: osx
       env:
         - PODSPEC=1
-    - os: linux
-      sudo: required
-      dist: trusty
     - os: linux
       sudo: required
       dist: trusty
