+++ Package.swift
@@ -1,47 +1,40 @@
-// swift-tools-version:3.1
+// swift-tools-version:4.0
 
-import Foundation
 import PackageDescription
 
-var isTesting: Bool {
-  guard let value = ProcessInfo.processInfo.environment["SWIFT_PACKAGE_TEST_Quick"] else { return false }
-  return NSString(string: value).boolValue
-}
-
-var package = Package(
+let package = Package(
     name: "Quick",
+    products: [
+        .library(name: "Quick", targets: ["Quick"]),
+    ],
+    dependencies: [
+        .package(url: "https://github.com/Quick/Nimble.git", from: "7.0.1"),
+    ],
     targets: {
+        var targets: [Target] = [
+            .testTarget(
+                name: "QuickTests",
+                dependencies: [ "Quick", "Nimble" ],
+                exclude: [
+                    "QuickAfterSuiteTests/AfterSuiteTests+ObjC.m",
+                    "QuickFocusedTests/FocusedTests+ObjC.m",
+                    "QuickTests/FunctionalTests/ObjC",
+                    "QuickTests/Helpers",
+                    "QuickTests/QuickConfigurationTests.m",
+                ]
+            ),
+        ]
 #if os(macOS)
-        return [
-            Target(name: "QuickSpecBase"),
-            Target(name: "Quick", dependencies: [ "QuickSpecBase" ]),
-            Target(name: "QuickTests", dependencies: [ "Quick" ]),
-        ]
+        targets.append(contentsOf: [
+            .target(name: "QuickSpecBase", dependencies: []),
+            .target(name: "Quick", dependencies: [ "QuickSpecBase" ]),
+        ])
 #else
-        return [
-            Target(name: "Quick"),
-            Target(name: "QuickTests", dependencies: [ "Quick" ]),
-        ]
+        targets.append(contentsOf: [
+            .target(name: "Quick", dependencies: []),
+        ])
 #endif
+        return targets
     }(),
-    exclude: {
-        var excludes = [
-            "Sources/QuickObjectiveC",
-            "Tests/QuickTests/QuickAfterSuiteTests/AfterSuiteTests+ObjC.m",
-            "Tests/QuickTests/QuickFocusedTests/FocusedTests+ObjC.m",
-            "Tests/QuickTests/QuickTests/FunctionalTests/ObjC",
-            "Tests/QuickTests/QuickTests/Helpers",
-            "Tests/QuickTests/QuickTests/QuickConfigurationTests.m",
-        ]
-#if !os(macOS)
-        excludes.append("Sources/QuickSpecBase")
-#endif
-        return excludes
-    }()
+    swiftLanguageVersions: [4]
 )
-
-if isTesting {
-  package.dependencies.append(contentsOf: [
-    .Package(url: "https://github.com/Quick/Nimble.git", majorVersion: 7),
-  ])
-}
