+++ Sources/Quick/World.swift
@@ -12,15 +12,9 @@ public typealias SharedExampleContext = () -> [String: Any]
 */
 public typealias SharedExampleClosure = (@escaping SharedExampleContext) -> Void
 
-// `#if swift(>=3.2) && (os(macOS) || os(iOS) || os(tvOS) || os(watchOS)) && !SWIFT_PACKAGE`
-// does not work as expected.
-#if swift(>=3.2)
-    #if (os(macOS) || os(iOS) || os(tvOS) || os(watchOS)) && !SWIFT_PACKAGE
-    @objcMembers
-    internal class _WorldBase: NSObject {}
-    #else
-    internal class _WorldBase: NSObject {}
-    #endif
+#if (os(macOS) || os(iOS) || os(tvOS) || os(watchOS)) && !SWIFT_PACKAGE
+@objcMembers
+internal class _WorldBase: NSObject {}
 #else
 internal class _WorldBase: NSObject {}
 #endif
